Event messages not printing
After further discussion, it was revealed that the event printouts would stop
coming out if a new event was created.  Found that the constructor of the
event was clearing out the static info_msg flag all events used to test
for printing.  Removed the assignment in the constructor and set the
initial value of the info_msg flag to false at the declaration point.

Fixes #38