remove broken/unused VarsWidget::updateSelection
This was using the wrong PlotModel indexes. I decided to
remove the functionality instead of fixing it.

When it worked, it synced up the select of a page in the
book view with the variable in the vars widget.

I don't really think that was all that helpful and possibly
confusing.