SWIG files in model libraries not being init'ed
Found that we need to add python files to lists of expected modules
before testing if they are being excluded.  Moved the check to
exclude to after adding the files to the list.

This may have uninteded consequences.  Might have to revisit using
different mechanism to exclude files and mark files as expected from
libraries.

refs #74