configure script reports missing motif as error.
So I found that I didn't write many of the autoconf rules correctly including
the one searching for motif.  I also messed up the checks for bison, flex, and
perl.  I added new rules to search for Xwindows headers and libraries.  The
motif rule is only executed if we find fermi-ware where it is used.

refs #119