dev: introduce BookIdxView and BookIdxView children for PlotView
This is a non-working version of snap on a big dev branch!  I wanted
to be able to go back to incremental checkins so just checked in
a lot of dev changes all at once. I decided it was worth the risk
of mixing in lots dev changes with new code.