Data products not plotting the last recorded data point #162
When reading a data stream to plot the end of stream flag is being set
as the last point is read.  This was causing several layers of calls
not to save the last point.  I went ahead and saved the result of the
get_next_point calls whether the end of stream flag is set or not and now
we don't lose the last point.