first cut at x version of legend
The pdf legend will be in a different commit.

The legend is a first cut because it is kind of an art on
how to draw it.  For instance, it shouldn't be drawn
if there are 1000 curves.  I don't draw it if it is an error
curve.  It would look funny if a page with 6 plots had
the same legend foreach plot.  So, this will morph with user
input.