runs subtitle sometimes chopped in GUI
Tim reported this.  When there is more than 1 run, the
title is multilined. With multiple lines,
the bottom line can get cut.  To eliminate this, I decided to
combine the lines into a single line.  This looks better and is the
predominant case.  It will fail only when the RUN_ names are very long.