formatting with %g can lose needed accuracy with live xy and flatline label
For example, formatting x=123456789.123456789 with %.9g
will format x to 123456789, sample program:

    // t.c
    #include <stdio.h>

    int main()
    {
        double x = 123456789.123456789;
        printf("x=%.9g\n", x);
        return 0;
    }

    % cc -o t t.c
    % ./t
    % x=123456789

That is because %.9g means to use 9 digits overall. %.9lf means use 9
digits after the decimal point (billionths).

In most cases, %g is better since it is more readable.  Use %.9g
unless the difference between the number formatted with %.9g and the
actual value is > 1.0e-9.  In the case that there is an error > 1.0e-9,
use %.9lf.