vertically align plots on a page
Made it so all plots on a page use the same space
y tic number labels.  This vertically aligns all the
plots on a page.

There still is one issue.  If the units change, the
y tic label numbers change and may cause the plot
to use different space which misaligns plots.  I can
force a redraw on all plots but thought I'd wait and
see if users find this necessary.