Unit test for SIM_test_dr failing. #411
This tst was only run if the HOST_CPU of the run matched one
of the pre-existing files.  So it was only tested on SL6.  Found
that the udunits change never made it to these log files.  I
updated the units and introduced only one copy of the files that
all platforms will test against.