big: initial commit of DP tables
This is a start.  I wrote my own table view since a standard QTableView
didn't work well as a view of a BookModel. The table view is called
a "BookTableView".  Hopefully, if this isn't enough, I will be
able to add missing functionality as the need arises.  It is fast.