mousewheel to change units very slow for monte carlo
The best optimization for this turned out to be
blocking signals when changing all units for each
curve.  This especially helped monte carlo
when viewing multiple pages of plots.

I had hoped to further optimize by updating the pixmap
only when necessary but didn't see an easy way to do that.