koviz barfs on run title in session file
The syntax for a run in the session file is:
  RUN: <run_dir> [run_title]

This fix only parses the run_title i.e. koviz will
not die if the run_title is specified.  The run_title is not
used.  I will make a new issue if users ask for the run_title
to be used for the tabs on the koviz GUI (or something similar
like using it for a title on pdfs).