rewrite -dp2trk commandline option
This supports -start/-stop options so that you can
extract a curve between a start and stop time.  It
also supports the -shift option so that you can
extract a curve and shift it.

The old way only worked for data that began at time 0.0.  This
method works for any time.

The old way only worked for data that had a uniform frequency.
This method works for variable frequency data.

This method is also much more simple.  I was able to delete
a lot of code from the TimeStamps class that was complex
and limited.

This method uses the same technique that BookTableView uses.