legend not being drawn as expected on qt 4.8
Instead of 3 rows, the legend is sometimes 2 rows.  Turns out
that the space between legend entries is actually negative
for a font size of 8.  Instead of using 0 for the space
between row entries, I use the font metric's leading() value.

I'm still unsure if this is the way I want to draw legends.
There is a catch22 between drawing in the available space
and drawing it based on the number of rows.