remove start/stop time from DataModel
I did this to simplify the DataModel i.e.
the concrete classes to come, like
CsvDataModel and Hdf5DataModel will not
need to handle -start/stop internally.

This affected the Snap real-time analysis.
You can't use the -start/stop time with the
-rt option.

The reason I had the -start/stop inside
the TrickModel to begin with was to
speed things up with large datasets.