native csv plotting
Prior to this, you had to convert csv to trk
in order to plot.  This is the first data format
to implement the DataModel interface and
ModelIterator.  It works well!

I tested a csv with 1100 columns.  I also
tested a 5GB csv file.  It took about 2 minutes
to load the csv file but was very fast once
loaded.