better top of the window spike selection for livecoord
Spikes at the top of the curves window were hard to select.
Sometimes it was necessary to pan the whole curve down in
order to hover over the spike.

To fix this, I made it so the spikes at the top of the
curves window are always selected if the mouse y-coord
is at the top 12.5% of the window.