plots with multiple curves with mixed units mishandled
There were a couple issues when a plot had
multiple curves with mixed units.

The first issue was that the y-axis {units} label
was picking up the first curve's units.  This
was very misleading.

The second issue was that error/compare plots needed
to autoconvert units if they were compatible yet
different e.g. if curve1 was {m} and curve2 was {cm}.

The third issue was that DP files with no unit
specs were not picking up the underlying
curve units.