empty plot title showing gridline when it should not
An empty plot title was showing a gridline when it
should not have.  The calculations for the viewport()
were giving unexpected results.  The
viewport()->rect().top and viewport()->rect().bottom
were both zero, however the viewport()->height() was
1 pixel.  So, before the fix, dv=viewport.bot-viewport.top was 0.
I changed it to dv=viewport.height.

I noticed that the plot title had
better calculation for coordToPixel transform