Disable setting TCP_NODELAY on UDP sockets
The TCP_NODELAY socket option disables Nagle's algorithm on TCP sockets,
improving latency at the cost of throughput. This option is
TCP-specific, and it is an error to use it on any other kind of socket,
including UDP. It is also unnecessary, since UDP is datagram-based and
does not buffer data to be transmitted.