Delete error path after using in _paintMarkers method
I noticed that the CurvesView::_paintMarkers method
was grabbing an error path using _bookModel->createCurvesErrorPath().
createCurvesErrorPath creates a painter path and expects the
client to delete it... so delete it after using it.

I need to cache error paths rather than creating on the fly.
Right now, each time a marker is drawn on an error path,
the entire painter path has to be regenerated.  I'm surprised
it runs as fast as it does with all that recreation.