DP tree has dates the cover up first column
I think there is a Qt bug that will not hide
columns even when the model is initialized
and the column is set to hidden. Anyway,
I guessed that the dp filter proxy model
might have something to do with it.  Returning
false inside  DPFilterProxyModel::filterAcceptsColumn
for columns > 0 seems to hide the date column.