Python init code does not work when using SWIG 4.0 #760
Looks like the way modules are loaded changed a little with SWIG 4.0. So I
changed the order we include everything in __init__.py.  We include all
of the c++ python modules first now instead of having them read in from
the python code itself.  Basically we include the files like this in this order

import _m<hash>
from m<hash> import *