Reimplement plot layout
This is big.  Instead of using the grid layout
to stretch the child widgets of a plot,
I had each widget report a size policy and
had each widget give a good size hint.

Now, hopefully, this will pave the way for
implementing the plot ratio feature.  This
is independent and is a step in that direction.

I tested this, but it's on my own machine.
If this breaks anybody, it's not the plot
ratio part, but me redoing the layout.