Introduce plot layout geometry manager
This is big.  Instead of using Qt's grid
layout class, I implemented my own.  This
is in preparation for square plots.

I spent a good deal of time trying to make
the grid layout work, but just could not
get it to behave properly.  I did this at
the advice of articles I read online.
Turns out that the PlotLayout class
was really easy to implement.  I wish I'd
done it all along.  Now that I have this,
I'm guessing I could consolodate the pdf
plotting with the X11 plots.