fix: TPROCTE_HOME Being Used In Place Of TPRO_HOME
When compiling Trick with the --with-tpro option for use of a real-time clock, the TPROCTE_HOME variable was being assigned the contents of the withval variable instead of being assigned to the TPRO_HOME variable. The TPROCTE_HOME variable was never used and the TPRO_HOME variable will always be an empty string. Thus, the configuration file will always check for only /h/tpro.h instead of <user-provided-path-to-tpro>/h/tpro.h