Do not throw out points with rt option
There was a point where the sims we were looking at with
TS21 had large spikes to ignore at the beginning of the run.
I decided to not throw out points any longer.  If there
are spikes at the beginning of the sim, they will show up.

Another issues was that spikes could be erroneously large. I
threw those out too.  Now, if there are spikes like that, I
need to get with the Trick guys and find out why.  I'm
guessing in Trick17 that issue does not exist.