Trick17 real-time iterator should use y value
For some reason, I was using the x parameter of
the iterator as runtime.  This caused an issue in Trick17
since, I think, has x set to time and y is runtime.

I tested the old Trick13 rt option and it->y() worked, so
this change should be fine for older versions.