Get configure with no java and no X to work #735
Found if --without-x was specified, that automatically caused an
error condition in the configure script.  Fixed that and only
tested for X windows headers and libs if X is included.  Excluded
a couple of directories from compiling if X is not included.

Didn't seem to be a problem with java.