Support external program units for input and output
Here is an example DP file with units specifications:

    External Program Test

    PROGRAM: "/home/rerun_data/test/ext_programs/myprogram"
      in: "ball.state.out.position[0] {cm}"
          "ball.state.out.position[1] {cm}"
     out: "mag {cm}"
          "sum {cm}"

    PLOTS:

      Page 1: "External Program Test"

        Plot 1: ""
          Y_axis_label: "Position"
          X_variable: "sys.exec.out.time"
          Y_variable: "ball.state.out.position[0]"
          Y_variable: "ball.state.out.position[1]"

        Plot 2: ""
          Y_axis_label: "Magnitude"
          X_variable: "sys.exec.out.time"
          Y_variable: "mag"
          UNITS: "m"

Note that one can specify the units on program
inputs and outputs as well as the PLOT units.

If the program input var unit is set, it
must be in the same unit family as logged.
The input will be converted before being
fed to the external program.  The program
output var units tells koviz what the
external program produces.  If plotting
the external program output var,
the PLOT units can be set as well for
the displaying the curve in the final
desired units.