Launch multiple user scripts and map to keystrokes
Example:
    % koviz RUN_* -script "&myconverter.py arg1, &vlc_wrapper.py"

Use an ampersand in front of the key to map to.  To launch by
keystroke, Alt+s followed by mapped key.  In the example, to launch
the vlc_wrappyer.py, you would press alt+s,v.

The user defined scripts will be on the main menu bar under
"Scripts".