Reimplement Runs class for RUN:var syntax
In order to support the RUN:var syntax
it was necessary to rethink how the Runs
class creates CurveModels.

This commit both simplifies the code and
makes for a more flexible approach.

This commit does not implement the RUN:var
capability, but will make it much more
easy to implement - so, consider this
stage 1.