X var defaulting to time if xvar not found
I found that my last checkin has a bug.  I wasn't checking
the timeName properly which was defaulting x to time if
x could not be found!  I found this while running one of the
tests in the test_maps.  The test was passing when it should
have been failing.