Sometimes QPainterPath::lineTo(x,y) does not add element
When points are very close to each other, for instance:
   p0=(1.0e-15,3.141592)
   p1=(2.0e-15,3.141592)
the lineTo(x,y) method will not add the point to the path.
I found this with some data from Jeff Bell (ES).

To get around this, I add a point whose distance from the
previous point is far enough to where Qt will add it, then
I change the position to what it is supposed to be.
That's the hack.