dp2trk and dp2csv options not working for small time epsilon
The TimeStamps class had a hard-coded value of 1.0e-9 for
dealing with timestamp comparison.  That was too large for
Jeff's dataset where timestamps were on the order of 1.0e-15.

I made the epsilon a static member of the TimeStamps class,
and set the value to 5.391245e-44 (Planck's time).  That was
an arbitrary choice!  The rational is that we are dealing
with sim time and sims that represent reality and don't think
any algorithms will need to deal with time epsilons that small.
We'll see!