Use best timestamp match for error plots
Prior to this, the error presentation
algorithm used the first timestamp it found
within the time match tolerance for the
delta.  This change picks the closest
timestamps within the time match tolerance.

This is a much better way because the
other way gave a different response
foreach time match tolerance you used!
That lead to misleading error plots.