Add -video and -videoOffset commandline options
The -video option takes a path to an *.mp4 to load.
The -videoOffset is a time offset that one can set
so that the video time can be adjusted to the data
time.  Another way to sync the video to the data
is by using the -shift option.  Eventually,
with multiple videos, each video will have its
own offset.  This is a step towards that.