Annotate empty plots and curves
Instead of just giving an empty plot,
label the plot "Empty" if there is
an empty plot.

If an empty curve is plotted along with
a curve with data, an "Empty" label
will still appear in the plot.

If an error curve has no data, for example
when the timestamps do not line up, the
error plot will be annotated with "Empty".