Mixing vars with different units when alt selecting vars broken
I had a faulty check that wasn't allowing a user to alt select
different vars with different units onto the same plot.  The
check was meant to make sure one didn't have different units
on the same var over a set of RUNs, but the check was too
strict, so I took it out.