Ubuntu 20.04, Python 3.8, and improved python configuration in configure.ac (#994) closes #991 closes #992 closes #993
see #991, #992, and #993.
To support Ubuntu 20.04 we must support new versions of python and gtest. Python has changed the way it provided libraries using python-config, so adjustments had to be made. I've completely ditched the built in autoconf function for finding python, because it adds a bunch of extra flags we don't need for Trick.

Now when we look for python with configure, it will search for python in this order:
python${PYTHON_VERSION}
python
python3

For python-config, it will search in this order:
python<major_version.minor_version>-config
python<major_version>-config
python${PYTHON_VERSION}-config
python-config
python3-config

Instead of python-config --ldflags, we are using python-config --libs
This only gets the flags we need.

For python >= 3.8 we use
python-config --libs --embed
this way it will add -lpython to the flags