Stylize curves by grouping RUNs by ranges and regexps
This commit adds the ability to group based on RUN
ranges and regular expressions.  For example:

% koviz RUN* -l1 "Initial RUN" -c1 pink -ls1 thick_line \
             -l2 "First Block" -c2 blue -g2 1,10 \
             -l3 "Last Run" -c3 red -g3 99,99 -ls3 thick_line \
             -t1 "Koviz Groups Example" -t2 "Spring Position" \
             -bg black -fg "#ededed"

The above will color the first curve pink, group color
RUN_00001-RUN_00010 blue, and will color the 99th RUN red.

To group based on name, use regexps, for example:

% koviz RUN* -l1 "Stable" -c1 blue -g1 "_S_" \
             -l1 "Unstable" -c2 red -g2 "_U_"

The above says, for RUNs with the string "_S_", color the
curves blue and label "Stable".  Similarly, color RUNs
that contain "_U_" in the RUN name as red, and label "Unstable".