Cleanup PlotMainWindow class
Instead of building the PlotBookModel
inside PlotMainWindow, it is built
in main().  This makes it so all the
commandline options don't have to
be passed to PlotMainWindow as args.
The options are added to the PlotBookModel
before being passed to PlotMainWindow.

Instead of three PlotMainWindows in main(),
there is only a single call.

This will make it much cleaner to add
new options.  Options will go directly
into the PlotBookModel and can
be accessed there.