Press escape to reset curves view
After zooming into a plot, one can
hit the right mouse button to go
to the original unzoomed view.

This commit adds hitting the <ESC>
key to do the same.

This will be useful for people who
do not have a middle mouse button
and bind zoom to the right mouse
and need to reset the view via
a key.