Support any number of plots on a page
Prior to this commit, the maximum number
of plots on a page was 7.  Historically,
seven came from the 7 jointed SSRMS
(station robotic arm).

I tried 100 plots on a page, but it
gets squishy.  I also tried a monte
carlo with 36 plots on a page.  This
was surprisingly good.  One can get
a good sense of the data with the
thumbnail plots.

I kept the manual layout with margins
for 1-7 plots.  For 8+ plots on a
page, I eliminated the margin and
autolayout the plots in a grid that is
as close to a square that is possible.

I think the layout could use a little
work with the y-axis label being too
close to the the adjacent plot and
not close enough to the y tic labels -
but overall this is at least a good
start.