Start zoom in plot title should work
I've wanted to fix this forever.  Before this
commit, starting a plot zoom in the plot
title did not rubberband.  With smaller plots or
data that maxes near the plot title, this
was annoying and unintuitive.

The issue was that the plot title is a
separate widget from the curves widget. The
rubberband logic was only in the curves widget.

There is a similar issue with scaling.  Right
now if one begins a scale by starting the mouse
drag in the plot tics, nothing happens.  I should
fix that too but it is less of an issue.  The
rubberband plot title issue was always biting me.