Video window jumping around with monte carlo
I hadn't seen this issue because I put
my video widget on one of my monitors
at full screen and keep it that way.
Tim was moving the video window to
a desired location then selecting
another RUN - which would cause the
video widget to hop to the original
location found when koviz last shutdown
via the settings file.