use default repo name in tutorial introduction
Novice linux/trick users don't catch that "trick" and "trick.git" are incompatible, because they don't understand that the file extension is a part of the directory name. For this reason they frequently leave the ".git" in their shell environment and they get stuck there for a while. While it is important that trick users have an understanding of how their shell uses the PATH variable, I do not think this little section at the bottom of the intro should have to be debugged so frequently.