Add slinch unit
This is for Marc J.

The slinch is a unit of mass invented by NASA.
The unit is part of a system based on the pound
of force and the inch. One slinch is the mass
accelerated at one inch per second per second by a
force of one pound; thus the slinch equals
exactly 12 slugs or about 386.088 pounds (175.1268 kilograms).
The word is a contraction of slug-inch. In the U.S. military
aircraft industry, this unit is sometimes called a mug.
It has also been called a snail.