Consolidate curve creation code
This is a left over TODO from the
drag-n-drop introduction.

The curve creation code needs the
monte inputs model to map runID to color.
I ended up passing the monte model
to the book model for curve creation.

This is all for sorting by monte
inputs to see color.  If the curves
"rainbow" there is a direct correlation
between the sorted input and associated
curves.