Drag selected curve onto another curve
This commit adds the ability to drag
a curve as well as drag a curve onto
another curve.

To drag a curve, first select
the curve, then hold the ctrl key and
drag with the left mouse.

To "snap" a curve onto another curve,
first select the curve, then with
alt+ctrl keys pressed, drag the initial point
of the dragged curve onto the
initial point of the other curve.
The curve will snap onto the other
curve when they get close to one another.