big: Toggle Between Time and Frequency Domain
Keypress 'f' on a plot will toggle to the
frequency domain.  Press 'f' again to go
back to time domain.

One may zoom in the Time domain, then toggle
to the frequency domain to see the zoomed
area in the frequency domain.

The xaxis in the "frequency domain" is the
fft frequency.  The xaxis contains all the
fft - so there is typically a mirror image
of the data centered at N/2.

The yaxis in the "frequency domain" is the
magnitude of the fft's complex vector for
the frequency on the xaxis.