Introduce interactive Butterworth filtering
To apply Butterworth filtering to a plot,
keypress 'b'.  A slider and entry box will
appear on the plot to set the frequency.

Koviz will automatically set and enforce
the valid frequency range in the slider and entry.

Koviz will also ensure the data's sample
rate is constant.

A descriptive popup will appear if there is any issue
with the data that would inhibit the filter from
running as it should.