Live coord off when x not time and using logscale
This issue was apparent when hovering over
a peak in logscale (say fft) when x was something
other than time - e.g. cte time.  Koviz would
draw the marker next to the peak rather than
the peak itself.

When the painter path x is not time, koviz
has to use the curve model (which has time)
rather than the painter path to find the point
on the path with the given time.

Normally, the painter path and curvemodel have the
same number of elements, so the time index in the
curvemodel is the same as the painter path.
When x is in logscale, zeroes get culled from
the painter path, so the indexes are not the same.

This commit fixes this odd case.