Support FFT when x is not sim time
Prior to this commit, the fft model used sim time
for the fft.  In practice, the x axis can be
something like cte time, so this commit adds
support when x is -any- time variable.

This commit adds a binary search method that
searches for a timestamp.  The search is more
complex than others - and came about
while testing the x-axis with fft frequencies.
There were subtle rounding issues, and
now wonder if the other classes that implement
the binary search should do it like CurveModelFFT::indexAtTime().