Use timedomain cache for time/freq domain toggle
When toggling from freq->time domain, an inverse fft
was used.  Instead of an inverse fft, decided to cache
the timedomain state and just return to it.  This makes
it so one can toggle from time->freq domain, hit 'f'
and get back to the last zoom in time domain, then
unzoom back to see all data in the time domain.
Bias/Scale is preserved as well.  Should make for
better flow.

The reason for using the inverse fft originally
was because I thought one might want to operate
in the freq domain (e.g. cut out frequencies)
and then return to the timedomain.  In practice,
this isn't something people do.