indexAtTime() not always picking best index for time
Due to subtle rounding issues, when time was not in
the data, the binary search algorithm wasn't always
picking the closest time.  This commit makes a final
pick between the top three points at the end of the
search.

The same algorithm is used by all classes, so had
to make changes across the board.  Really need to
put this algorithm in a base class that all could
inherit.  Unfortunately, it's not straightforward.