Give popup error msg when attempting to filter Hz data
Tim accidentally hit 'g' after switching to the frequency
domain with 'f'.  When in the frequency domain, xunits are
Hz.  Hitting 'g' or 'b' inadvertantly while in the freq
domain should give a nice popup like "Butterworth filter
expects xunits to be in seconds.  Try again".
Prior to this commit, there was a hard exit.  In general,
I really need to stop exiting and always give a nice
popup.