Derivatives!
This commit adds the ability to take the derivative
of a curve with respect to time.  Keypress 'd' to
take the derivative.  Units work with derivatives
as well.  One may take successive derivatives by
repeatedly pressing 'd'.

The derivative handles nans and successive points
on the curve with the same timestamp.

The derivative technique uses the average of
the slopes before and after each point.