Pdf plot borders are missing
I used a pen of width 0 for plot borders/tics.  Something
must have changed because the borders are now gone with
a width of zero.  Not sure what has changed.  A pen
width of zero is supposed to be "cosmetic" meaning there
should be a line of width 1 under any transformation.