Turn off plot legend if -legend is off
The -legend option is actually for the page legend; however,
it makes sense to tie -legend off for turning off page and
plot legends.

This commit makes it so if -legend is off and
-showPlotLegend is unset, then turn -showPlotLegend off.

If -legend is on and -showPlotLegend is unset, the plot
legend will not necessarily be enabled.  That's not
symmetrical but think it will work in practice.

This started because we had a plot legend that was
obscuring the plot, and we tried -legend off and still
got the legend - not seeing the -showPlotLegend option.

Anyhow, now -legend off will turn off the plot legend
unless -showPlotLegend is explicity set to yes.