Hide plot legend if all labels identical
There are cases where koviz was showing
plot legends where all the labels were
identical.  This, for example, could happen
when the DP file had two plots on a page
and koviz was called with two RUNs.
In this case, each plot would have a plot
legend with two identical labels.