Have synced video start at time zero rather than negative time
When syncing a video with a video offset, the video could roll
for a while in negative time until the video caught up with
the data. This commit makes it so the video begins when the data
begins rather than counting off in negative time.

The only issue with this is that one will not see the start
of the video, the video time will hop in time.  If the
start of the video needs to be viewed, it's still possible -
just paused the video and rewind to the very beginning.